Pull out WorkshopTypeOptions from workshop_form.jsx
